[Basic] Index Type
This pull request introduces the `Index` type to represent the index inside ElasticSearch.

Even though downstream consumers need to make their own implementations of the index that can be serialized to JSON, it's good to have a base class to start from.
There is a given structure on the index for ElasticSearch that can be enforced from our side to avoid invalid data being submitted to the cluster.